refactor: Change refresh folders function to MessagesActionsUseCase
