+++ app/src/main/java/com/infomaniak/mail/ui/MainViewModel.kt
@@ -821,7 +821,7 @@ class MainViewModel @Inject constructor(
         destinationFolderId: String? = null,
         callbacks: RefreshCallbacks? = null,
     ) = viewModelScope.launch(ioCoroutineContext) {
-        sharedUtils.refreshFolders(mailbox, messagesFoldersIds, destinationFolderId, currentFolderId, callbacks)
+        messagesActionsUseCase.refreshFolders(mailbox, messagesFoldersIds, destinationFolderId, currentFolderId, callbacks)
     }
 
     private fun onDownloadStart() {
